@@ -1,12 +1,13 @@
 {
   "name": "flarum-webpack-config",
-  "version": "0.1.0-beta.10",
+  "type": "module",
+  "version": "1.0.1",
   "description": "Webpack config for Flarum JavaScript compilation.",
-  "main": "index.js",
+  "main": "src/index.js",
   "author": "Flarum Team",
   "license": "MIT",
   "peerDependencies": {
-    "webpack": "^4.0.0"
+    "webpack": "^5.36.2"
   },
   "dependencies": {
     "@babel/core": "^7.14.0",
@@ -15,12 +16,24 @@
     "@babel/plugin-transform-object-assign": "^7.12.13",
     "@babel/plugin-transform-react-jsx": "^7.13.12",
     "@babel/plugin-transform-runtime": "^7.13.15",
-    "@babel/preset-env": "^7.14.0",
+    "@babel/preset-env": "^7.14.1",
     "@babel/preset-react": "^7.13.13",
     "@babel/runtime": "^7.14.0",
-    "babel-loader": "^8.2.2"
+    "babel-loader": "^8.2.2",
+    "loader-utils": "^1.4.0",
+    "schema-utils": "^3.0.0"
   },
   "devDependencies": {
-    "prettier": "^2.2.1"
+    "babel-jest": "^26.6.3",
+    "jest": "^26.6.3",
+    "memfs": "^3.2.2",
+    "prettier": "^2.2.1",
+    "webpack": "*"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
